% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBConnection.R
\name{connectToDatabase}
\alias{connectToDatabase}
\title{connectToDatabase}
\usage{
connectToDatabase()
}
\value{
A connection object to the Serratus database if successful,
NULL otherwise.
}
\description{
Function to connect to the Serratus database.
}
\examples{
conn <- connectToDatabase()

}
