% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLInterface.R
\name{constructQuery}
\alias{constructQuery}
\title{constructQuery}
\usage{
constructQuery(SQLtable = NULL, filters = NULL)
}
\arguments{
\item{SQLtable}{The name of the SQL table to query.}

\item{filters}{A named list of filters to apply to the query.}
}
\value{
A SQL query string.
}
\description{
Constructs a SQL query based on the provided filters.
}
\keyword{internal}
