% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLInterface.R
\name{constructSelect}
\alias{constructSelect}
\title{constructSelect}
\usage{
constructSelect(filters = NULL)
}
\arguments{
\item{filters}{A named list of filters to apply to the query.}
}
\value{
A SELECT clause string.
}
\description{
Constructs the SELECT clause of a SQL query.
}
\keyword{internal}
