% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLInterface.R
\name{constructWhere}
\alias{constructWhere}
\title{constructWhere}
\usage{
constructWhere(filters = NULL)
}
\arguments{
\item{filters}{A named list of filters to apply to the query.}
}
\value{
A WHERE clause string.
}
\description{
Constructs the WHERE clause of a SQL query.
}
\keyword{internal}
