% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLInterface.R
\name{executeQuery}
\alias{executeQuery}
\title{executeQuery}
\usage{
executeQuery(conn = NULL, query = NULL)
}
\arguments{
\item{conn}{A DBI connection object.}

\item{query}{A SQL query string.}
}
\value{
The result of the query.
}
\description{
Executes a SQL query on the provided connection.
}
\keyword{internal}
