% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseSQL.R
\name{getAllDataProcessed}
\alias{getAllDataProcessed}
\title{getAllDataProcessed}
\usage{
getAllDataProcessed(
  conn = NULL,
  librarySource = NA,
  runID = NA,
  biosampleID = NA,
  bioprojectID = NA
)
}
\arguments{
\item{conn}{A database connection object.}

\item{librarySource}{A character string specifying the library source taxa.}

\item{runID}{A vector of run IDs.}

\item{biosampleID}{A vector of biosample IDs.}

\item{bioprojectID}{A vector of bioproject IDs.}
}
\value{
A tibble containing the processed data.
}
\description{
Retrieves all data for the given query parameters processed by
Serratus (i.e. not just virus positive runs).
}
\examples{
conn <- connectToDatabase()
query <- list(conn = conn, librarySource = "Nematoda")
data <- getAllDataProcessed(query)
}
