% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityCalculations.R
\name{getAlphaDiversity}
\alias{getAlphaDiversity}
\title{getAlphaDiveristy}
\usage{
getAlphaDiversity(virome = NULL, mode = "shannon", bioSample = NULL)
}
\arguments{
\item{virome}{A virome object}

\item{mode}{A character vector of diversity metrics to calculate. Options:
"shannon", "simpson", "richness", "evenness". Default is "shannon".}

\item{bioSample}{Optional argument for filtering the virome by BioSample}
}
\value{
A tibble with columns for each diversity metric calculated.
}
\description{
A wrapper function for calculating various metrics of alpha
diversity for a virome object. Options to calculate Shannon diversity,
Simpson diversity, species richness, and Pielou evenness. Each bioSample
is treated as an independent sample for calculating these metrics. A more
detailed explanation in the context of SRA virome data is available in the
vignette.
}
\examples{
# Calculate Shannon diversity for all virome data
data("TylenchoideaVirome")
getAlphaDiversity(TylenchoideaVirome) # Shannon diversity by default

# Calculate Shannon diversity for a specific biosample
getAlphaDiversity(TylenchoideaVirome, bioSample = "SAMN25258043")

# Calculate Simpson diversity for all virome data
getAlphaDiversity(TylenchoideaVirome, mode = "simpson")

# Calculate all metrics for all virome data
getAlphaDiversity(TylenchoideaVirome, mode = c("shannon", "simpson",
                  "richness", "evenness"))

}
