% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityCalculations.R
\name{getAmpvisCounts}
\alias{getAmpvisCounts}
\title{getAmpvisData}
\usage{
getAmpvisCounts(virome, con)
}
\arguments{
\item{virome}{A virome object}

\item{con}{A database connection object (to Serratus SQL).}
}
\value{
A data.frame with OTUs as rows and samples as columns, with values
being the normalized number of reads mapping to each OTU in each sample.
}
\description{
Return an OTU-table and metadata table in ampvis format
(OTUs as rows, samples as columns), values are the normalized number of
reads mapping to each OTU in each sample. Metadata table is a subset of the
palm_virome table. Columns are:
\itemize{
\item bioSample: BioSample accession
\item bioProject: BioProject accession
\item source: Metadata annotation of the BioSample source species
NOT EXPORTED
}
}
\keyword{internal}
