% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viromeQueries.R
\name{getAvailablePhyla}
\alias{getAvailablePhyla}
\title{getAvailablePhyla}
\usage{
getAvailablePhyla(virome = NULL)
}
\arguments{
\item{virome}{A virome object}
}
\value{
A character vector of phyla
}
\description{
Returns a list of phyla that are present in the virome.
Mostly for use by the shiny frontend.
}
\examples{
con <- palmid::SerratusConnect()
virome <- getVirome(tax="Canis", con=con)
phyla <- getAvailablePhyla(virome = virome)

}
