% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityCalculations.R
\name{getBetaDiversity}
\alias{getBetaDiversity}
\title{getBetaDiversity}
\usage{
getBetaDiversity(virome)
}
\arguments{
\item{virome}{A virome object}
}
\value{
A tibble with rows for each bioSample present in the virome and a
column for each bioSample, with the Bray-curtis dissimilarity between the
two bioSamples.
}
\description{
Calculate pairwise Bray-curtis dissimilarity of a virome object
for all bioSamples present in the virome. Bray-curtis dissimilarity is
defined as sum(|x_ij - x_ik|) / sum(x_ij + x_ik), where where x_ij and
x_ik are the number of reads mapping to sotu i and bioSample j and k,
respectively.
}
