% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicPlots.R
\name{getDefaultViromePalette}
\alias{getDefaultViromePalette}
\title{getDefaultViromePalette}
\usage{
getDefaultViromePalette(numColors = 9)
}
\arguments{
\item{numColors}{The number of colors to return. Default is 9.}
}
\value{
A vector of colors.
}
\description{
Returns a default color palette for virome plots (Okabe-ito).
}
\examples{
getDefaultViromePalette()
getDefaultViromePalette(numColors = 5)

}
