% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityCalculations.R
\name{getDiversity}
\alias{getDiversity}
\title{getDiversity}
\usage{
getDiversity(virome, mode = "shannon")
}
\arguments{
\item{virome}{A virome object}

\item{mode}{Either "shannon" or "simpson"}

\item{con}{A database connection object (to Serratus SQL).}
}
\value{
A tibble with rows for each bioSample present in the virome and a
column for type of diversity calculated. Each bio_sample is treated as a
sample, and each sotu is treated as a species. Reads mapping to each sotu
are treated as the abundance of that species.
}
\description{
Calculate either Shannon or Simpson diversity of a virome object
Note: There are multiple metrics referred to in the literature as "Simpson
diveristy". This function calculates the D_0, which is the probability that
two randomly selected reads will be of the same sotu.
NOT EXPORTED
}
\keyword{internal}
