% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityCalculations.R
\name{getEvenness}
\alias{getEvenness}
\title{getEvenness}
\usage{
getEvenness(virome)
}
\arguments{
\item{virome}{A virome object}
}
\value{
A tibble with rows for each bioSample present in the virome and a
column for evenness.
}
\description{
Calculate Pielou evenness of a virome object.
Note: Pielou evenness is defined as the Shannon diversity divided by the
natural log of the richness. Evenness metrics have known issues as they
are largely dependent on species (in this case reads) abundance, which
can be variable between samples (due to sequencing depth). Some values
may be NaN if the richness is 1.
NOT EXPORTED
}
\keyword{internal}
