% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hostViralBurden.R
\name{getHostViralBurden}
\alias{getHostViralBurden}
\title{getHostViralBurden}
\usage{
getHostViralBurden(virome = NULL)
}
\arguments{
\item{virome}{A virome object}
}
\value{
A data frame with the total number of runs queried and the number
of runs that were virus positive for each source species.
}
\description{
Takes a virome object as input and re-queries Serratus to get
the incidence of viral RNA positive SRA runs out of the total number queried
for each source species.
}
\examples{
con <- palmid::SerratusConnect()
virome <- getVirome(tax = "Meloidogyne", con = con)
getHostViralBurden(virome = virome, con = con)
}
