% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityCalculations.R
\name{getRichness}
\alias{getRichness}
\title{getRichness}
\usage{
getRichness(virome)
}
\arguments{
\item{virome}{A virome object}
}
\value{
A tibble with rows for each bioSample present in the virome and a
column for richness.
}
\description{
Calculate species richness (i.e. number of unique viral
families) of a virome object.
NOT EXPORTED
}
\keyword{internal}
