% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDiversity.R
\name{getShannonDiversity}
\alias{getShannonDiversity}
\title{getShannonDiversity}
\usage{
getShannonDiversity(virome = NULL, sourceSpecies = NULL)
}
\arguments{
\item{virome}{A virome object}

\item{sourceSpecies}{Optional argument for filtering by source species}
}
\value{
A numeric value (Shannon Diversity Index)
}
\description{
Calculate Shannon diversity of a virome object
}
