% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDiversity.R
\name{getSimpsonDiversity}
\alias{getSimpsonDiversity}
\title{getSimpsonDiversity}
\usage{
getSimpsonDiversity(virome = NULL, sourceSpecies = NULL)
}
\arguments{
\item{virome}{A virome object}

\item{sourceSpecies}{Optional argument for filtering by source species}
}
\value{
A numeric value (Simpson Diversity Index)
}
\description{
Calculate Simpson diversity of a virome object
}
