% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityCalculations.R
\name{getSpeciesCounts}
\alias{getSpeciesCounts}
\title{getSpeciesCounts}
\usage{
getSpeciesCounts(virome = NULL, con = NULL)
}
\arguments{
\item{virome}{A virome object}

\item{con}{A database connection object (to Serratus SQL).}
}
\value{
A tibble with rows for each bioSample present in the virome and a
column for each sotu, with the normalized number of reads mapping to that
sotu.
}
\description{
Calculate the number of reads mapping to each sotu in a virome
object.
NOT EXPORTED
}
\keyword{internal}
