% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDiversity.R
\name{getSpeciesRichness}
\alias{getSpeciesRichness}
\title{getSpeciesRichness}
\usage{
getSpeciesRichness(virome = NULL, sourceSpecies = NULL)
}
\arguments{
\item{virome}{A virome object}

\item{sourceSpecies}{Optional argument for filtering by source species}
}
\value{
A numeric value (Species Richness)
}
\description{
Calculate species richness of a virome object
}
