% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viromeQueries.R
\name{getVirome}
\alias{getVirome}
\title{getVirome}
\usage{
getVirome(tax = NULL, sra = NULL, con = NULL)
}
\arguments{
\item{tax}{A taxonomic term}

\item{sra}{A character vector of SRA accessions}

\item{con}{A connection to the Serratus database}
}
\value{
A virome object
}
\description{
Returns a virome object given either:
\enumerate{
\item a taxonomic term and (optional) taxonomic level
\item a character vector of SRA accessions
Note: the broader the taxonomic search term, the longer the query will
take.
}
}
\examples{
con <- palmid::SerratusConnect()
getVirome(tax = "Meloidogyne", con = con)
getVirome(sra = c("SRR17756040", "SRR5942326"), con = con)

}
\references{
Wickham H, François R, Henry L, Müller K, Vaughan D (2023).
dplyr: A Grammar of Data Manipulation.
https://dplyr.tidyverse.org, https://github.com/tidyverse/dplyr.

A. Babaian and R. C. Edgar (2022),
Ribovirus classification by a polymerase barcode sequence. PeerJ

R. C. Edgar et al. (2021),
Petabase-scale sequence alignment catalyses viral discovery. Nature.
}
