% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viromeQueries.R
\name{getVirusTaxonomy}
\alias{getVirusTaxonomy}
\title{getVirusTaxonomy}
\usage{
getVirusTaxonomy(otu = NULL, con = NULL)
}
\arguments{
\item{otu}{A character vector of sOTUs}

\item{con}{A connection to the Serratus database}
}
\value{
A dataframe of taxonomic information for the specified virus
}
\description{
Return a dataframe with the tax_phylum for each sOTU in otu.
}
