% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidencePlot.R
\name{incidencePlot}
\alias{incidencePlot}
\title{incidencePlot}
\usage{
incidencePlot(virome, con)
}
\arguments{
\item{virome}{A virome object}

\item{con}{A database connection}
}
\value{
A plotly bar chart
}
\description{
Plot the incidence of virus positive sra runs against all
runs analyzed
}
