% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viromeQueries.R
\name{nameVecToRank}
\alias{nameVecToRank}
\title{nameVecToRank}
\usage{
nameVecToRank(names = NULL, taxRank = NULL)
}
\arguments{
\item{names}{A character vector of scientific names}

\item{taxRank}{A character vector of taxonomic ranks}
}
\value{
A character vector of taxons
}
\description{
Convert a vector of scientific names to the specified taxonomic
ranks for those names.
}
