% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palmPrevalence.R
\name{palmPrevalence}
\alias{palmPrevalence}
\title{palmPrevalence}
\usage{
palmPrevalence(virome)
}
\description{
For every sotu in a virome, plot the number of runs in which
it was found against the number of bioprojects
}
