% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palmPrevalence.R
\name{palmPrevalenceKernels}
\alias{palmPrevalenceKernels}
\title{palmPrevalenceKernels}
\usage{
palmPrevalenceKernels(virome)
}
\arguments{
\item{virome}{A virome object}
}
\value{
A plotly plot
}
\description{
For each of the four distributions present in the palmPrevalence
plot (SRA runs, bio_projects, mean coverage, and genbank identity), plot the
kernel density estimate.
}
\examples{
con <- palmid::SerratusConnect()
virome <- getVirome("Meloidogyne", con = con)
palmPrevalenceKernels(virome)

}
