% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perSpeciesHeatmap.R
\name{perSpeciesHeatmap}
\alias{perSpeciesHeatmap}
\title{perSpeciesHeatmap}
\usage{
perSpeciesHeatmap(virome)
}
\arguments{
\item{virome}{A virome object}
}
\value{
A plotly heatmap
}
\description{
creates a heatmap of the number of palmids aligned to a given
family for each species in the genus-level virome.
}
