% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityCalculations.R
\name{plotBetaDiversity}
\alias{plotBetaDiversity}
\title{plotBetaDiversity}
\usage{
plotBetaDiversity(virome = NULL)
}
\arguments{
\item{virome}{A virome object}
}
\description{
Automatically plots beta diversity for a virome object using
ampvis2. This is the plot used for the shiny frontend, and uses Bray-Curtis
dissimilarity with PCOA. There are other options for plotting beta diversity,
see https://kasperskytte.github.io/ampvis2/articles/ampvis2.html#ordination.
}
\examples{
con <- palmid::SerratusConnect()
virome <- getVirome(tax = "Salidae", con = con)
plotBetaDiversity(virome)

}
