% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVirome.R
\name{plottingHelper}
\alias{plottingHelper}
\title{plottingHelper}
\usage{
plottingHelper(virome = NULL, phylumFilter = NULL, colors = NULL)
}
\value{
Filtered virome object with an added column for colors assigned to
phyla.
}
\description{
Helper function for the pie chart and sankey diagram functions.
Does some sanity checks on user's input filters the data based on the
provided phylumFilter.
NOT EXPORTED
}
