% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVirome.R
\name{plotVirome}
\alias{plotVirome}
\title{plotVirome}
\usage{
plotVirome(virome)
}
\arguments{
\item{virome}{A virome object}
}
\value{
A plotly pie chart
}
\description{
creates a pie chart of the number of palmids aligned to a given
family for each species in the genus-level virome.
}
