% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexPlots.R
\name{plotViromeNetwork}
\alias{plotViromeNetwork}
\title{plotViromeNetwork}
\usage{
plotViromeNetwork(
  virome = NULL,
  threshold = 0.01,
  splitVariable = "bioprojectID"
)
}
\arguments{
\item{virome}{A data frame containing columns 'sOTU,' 'bioprojectID,' and
'normalizedNodeCoverage'}

\item{threshold}{A numeric value between 0 and 1 that determines the threshold for
the Jaccard similarity of sOTUs between bioprojects. Default is 0.0.}

\item{splitVariable}{A character string that determines how the virome data should be
split. Default is 'bioprojectID.'}
}
\value{
A networkD3 plot
}
\description{
plotViromeNetwork
}
\examples{
virome <- retrieveOTUs(bioprojectID = NA, librarySource = "Tylenchoidea", 
normalizedNodeCoverage = NA)
plotViromeNetwork(virome = virome, threshold = 0.01, splitVariable = "bioprojectID")

}
