% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVirome.R
\name{plotViromePie}
\alias{plotViromePie}
\title{plotViromePie}
\usage{
plotViromePie(virome = NULL, phylumFilter = NULL, colors = NULL)
}
\arguments{
\item{virome}{A virome object}

\item{phylumFilter}{Optional argument for filtering the virome by phylum.
Default is NULL and will not filter the data. Must be a character vector.}

\item{colors}{Optional argument for specifying colors for the pie chart.
Default is NULL and will use the Okabe-Ito palette if there are 9 or fewer
phyla. Otherwise, uses RColorBrewer Set3. Must be a character vector of hex
colors. length(colors) == length(unique(virome$tax_phylum))}
}
\value{
A plotly pie chart
}
\description{
Returns a plotly pie chart of predicted phyla for unique sOTUs
in the virome object.
}
\examples{
con <- palmid::SerratusConnect()
virome <- getVirome(tax = "Meloidogyne", con = con)
plotViromePie(virome = virome,
           phylumFilter = c("Pisuviricota", "Kitrinoviricota"),
           colors = c("#E69F00", "#56B4E9"))

}
