% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hostViralBurden.R
\name{plotVirusPositive}
\alias{plotVirusPositive}
\title{plotVirusPositive}
\usage{
plotVirusPositive(vbDF = NULL)
}
\value{
A plotly bar chart
}
\description{
Plots the proportion of virus positive runs for each taxon
}
\examples{
vbDF <- getHostViralBurden(virome = virome)
plotVirusPositive(vbDF)
}
