% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLInterface.R
\name{processInequalityArgs}
\alias{processInequalityArgs}
\title{processInequalityArgs}
\usage{
processInequalityArgs(requestedInequalities = NULL)
}
\arguments{
\item{requestedInequalities}{A named list of inequality parameters.}
}
\value{
An inequality clause string.
}
\description{
Processes the inequality parameters in the WHERE clause.
}
\keyword{internal}
