% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseSQL.R
\name{processUserArgs}
\alias{processUserArgs}
\title{processUserArgs}
\usage{
processUserArgs(args = NULL)
}
\arguments{
\item{args}{A named list of arguments to be processed. The names should
correspond to the supported parameter names, and the values are the arguments'
values.}
}
\value{
A list containing two elements: \code{columns}, a named list of arguments
mapped to their SQL column names for database queries, and \code{specialArgs},
a list of arguments that require special handling.
}
\description{
Processes user-provided arguments, ensuring they are valid and
categorizing them into columns (for database selection) and special arguments
(requiring further processing).
}
\examples{
processUserArgs(list(librarySource = 'Homo sapeins', 
STAT = c('Toxoplasma gondii'), 
minKmerPerc = 0.5))

}
\keyword{internal}
