% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLInterface.R
\name{processWhereArgs}
\alias{processWhereArgs}
\title{processWhereArgs}
\usage{
processWhereArgs(whereArgs = NULL, index = 1)
}
\arguments{
\item{whereArgs}{A named list of filters to apply to the query.}

\item{index}{The index of the current filter.}
}
\value{
A WHERE clause string.
}
\description{
Proccesses the WHERE clause arguments. Helper for
constructWhere.
}
\keyword{internal}
