% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseSQL.R
\name{retrieveTaxIDs}
\alias{retrieveTaxIDs}
\title{retrieveTaxIDs}
\usage{
retrieveTaxIDs(taxNames = NULL)
}
\arguments{
\item{taxNames}{A character vector of taxonomic names.}
}
\value{
A named list of taxonomic IDs, if successful. Otherwise throws an
error.
}
\description{
Retrieves the NCBI taxonomic IDs for a given set of taxonomic
names.
}
\examples{
taxNames <- c("Homo sapiens")
retrieveTaxIDs(taxNames)

}
\keyword{internal}
