% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseSQL.R
\name{Rnames2SQLnames}
\alias{Rnames2SQLnames}
\title{Rnames2SQLnames}
\usage{
Rnames2SQLnames(Rnames = NULL)
}
\arguments{
\item{Rnames}{A character vector of R parameter names.}
}
\value{
A character vector of SQL column names.
}
\description{
Translates R parameter names to SQL column names.
}
\examples{
Rnames2SQLnames(c("runID", "biosampleID", "bioprojectID"))

}
\keyword{internal}
