% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viromeQueries.R
\name{scientificNametoTaxID}
\alias{scientificNametoTaxID}
\title{scientificNametoTaxID}
\usage{
scientificNametoTaxID(name)
}
\arguments{
\item{name}{A character vector of scientific names}
}
\value{
A character vector of taxonomic IDs
}
\description{
Convert a scientific name to a taxonomic ID
}
