% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseSQL.R
\name{specialFilter}
\alias{specialFilter}
\title{specialFilter}
\usage{
specialFilter(conn = NULL, virome = NULL, filters = NULL)
}
\arguments{
\item{conn}{A connection to the Serratus SQL database.}

\item{virome}{A dataframe containing the current virome data.}

\item{filters}{A list of filters to apply to the virome data.}
}
\value{
A dataframe containing the virome data with the special filters
applied.
}
\description{
A function to handle special filters that require additional
data from the SQL not present in the palm_virome table.
}
\keyword{internal}
