% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseSQL.R
\name{SQLnames2Rnames}
\alias{SQLnames2Rnames}
\title{SQLNames2Rnames}
\usage{
SQLnames2Rnames(sqlNames = NULL)
}
\arguments{
\item{sqlNames}{A character vector of SQL column names.}
}
\value{
A character vector of R parameter names.
}
\description{
Translates SQL column names to R parameter names.
}
\examples{
SQLNames2Rnames(c("run", "biosample_id", "bioproject_id"))

}
\keyword{internal}
