% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseSQL.R
\name{taxHierarchyTranslate}
\alias{taxHierarchyTranslate}
\title{taxHierarchyTranslate}
\usage{
taxHierarchyTranslate(taxIDs = NULL, level = NULL)
}
\arguments{
\item{taxIDs}{A vector of taxonomic IDs (must be NCBI taxIDs).}

\item{level}{The level in the taxonomic hierarchy to translate the IDs to.}
}
\value{
A vector of taxonomic IDs at the specified level in the hierarchy.
}
\description{
Given a set of taxonomic IDs, translate them to a specified
level in the taxonomic hierarchy. Collects all child taxa.
}
\examples{
taxHierarchyTranslate(taxIDs = c(9606, 10090), level = "order")

}
\keyword{internal}
