% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVirome.R
\name{taxLookup}
\alias{taxLookup}
\title{taxLookup}
\usage{
taxLookup(tax, con)
}
\arguments{
\item{tax}{A taxon defined in NCBI taxonomy. Must be type char.}

\item{con}{A connection to the Serratus database}
}
\value{
A character vector of SRA accessions.
}
\description{
Return a list of all runs processed by Serratus that match
tax. Used internally by getVirome.
}
