% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TylenchoideaVirome}
\alias{TylenchoideaVirome}
\title{TylenchoideaVirome}
\format{
A data frame with 595 rows and 18 columns. Each row represents an
individual vOTU (viral operational taxonomic unit). Columns as follows:
\describe{
\item{run}{SRA run accession number.}
\item{scientific_name}{Metadata annotation for the library source species.}
\item{bio_sample}{SRA BioSample accession number.}
\item{bio_project}{Source BioProject accession number.}
\item{palm_id}{Serratus identifier for this palm print.}
\item{sotu}{Serratus identifier for the clustered sOTU (species-like
operational taxonomic unit).}
\item{nickname}{Serratus nickname for the sOTU.}
\item{gb_acc}{GenBank accession for the closest aligned sequence via BLAST}
\item{gb_pid}{Percent identity of the BLAST alignment.}
\item{gb_eval}{E-value for the BLAST alignment.}
\item{tax_species}{Taxonomic species annotation for the GenBank aligned
virus.}
\item{node}{Node number for the vOTU.}
\item{node_coverage}{Coverage of the vOTU in this library.}
\item{node_pid}{Node percent identity.}
\item{node_eval}{Node E-value.}
\item{node_qc}{Bool; whether the node passed quality control checks.}
\item{node_seq}{Detected palm sequence for this vOTU.}
}
\usage{
TylenchoideaVirome
}
\description{
Filtered Serratus SQL data with the member genus of the Tylenchoidea
superfamily.
}
\examples{
\dontrun {
TylenchoideaVirome
}

}
\references{
Edgar, R.C., Taylor, B., Lin, V. et al. Petabase-scale sequence alignment
catalyses viral discovery. Nature 602, 142–147 (2022).
https://doi.org/10.1038/s41586-021-04332-2
}
\keyword{datasets}
