% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viroMap.R
\name{viroMap}
\alias{viroMap}
\title{viroMap}
\usage{
viroMap(virome = NULL, minCov = 0)
}
\arguments{
\item{virome}{A tbl_df containing the virome data.}

\item{minCov}{The minimum coverage value to include in the heatmap.}
}
\description{
Plot a heatmap for the virome. Each column is a sample
and each row is an sOTU. The intensity of color for each cell is proportional
to the log of the node_coverage value for that sOTU in that sample. Samples
are grouped by source species. sOTUs are grouped by family. Note: the heatmap
might not display properly in the RStudio viewer and should be viewed in a
separate window.
}
\examples{
con <- palmid::SerratusConnect()
virome <- getVirome(tax = "Meloidogyne", con = con)
viroMap(virome = virome)

}
