% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityCalculations.R
\name{viromeToMetadataTable}
\alias{viromeToMetadataTable}
\title{viromeToMetadataTable}
\usage{
viromeToMetadataTable(virome = NULL)
}
\description{
Convert a virome object to the metadata table format specified
in ampvis2. First column specifies biosample id's, second column specifies
bioproject id's, third column is the scientific_name of the sample.
}
\examples{
con <- palmid::SerratusConnect()
virome <- getVirome(tax = "Salidae", con = con)
metadata <- viromeToMetadataTable(virome)
head(metadata)

}
