% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityCalculations.R
\name{viromeToOTUTable}
\alias{viromeToOTUTable}
\title{viromeToOTUTable}
\usage{
viromeToOTUTable(virome)
}
\arguments{
\item{virome}{A virome object}
}
\description{
Convert a virome object to the ampvis2 otutable format where
each row is a viral sOTU and each column is a biosample. The last two
columns provide the tax_species and tax_phylum for each sOTU.
}
\examples{
con <- palmid::SerratusConnect()
virome <- getVirome(tax = "Salidae", con = con)
otuTable <- viromeToOTUTable(virome)
head(otuTable)

}
