% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVirome.R
\name{drawVirusSankey}
\alias{drawVirusSankey}
\title{drawVirusSankey}
\usage{
drawVirusSankey(virome, phylumFilter = NULL, colors = NULL)
}
\arguments{
\item{virome}{A virome object}

\item{phylumFilter}{A character vector of viral phyla to filter by. Default
is NULL and will not filter the data.}

\item{colors}{A character vector of colors to use for the Sankey diagram.
Default is NULL and will use the Okabe-Ito palette if there are 9 or fewer
phyla. Otherwise, uses RColorBrewer Set3. Must be a character vector of hex
colors. length(colors) == length(unique(virome$tax_phylum))}
}
\value{
A plotly Sankey diagram of the virome object
}
\description{
Draws a Sankey diagram of the virome object with the following
columns (from left to right): predicted phylum, closest aligned species in
GenBank, Serratus sOTU, BioSample accession, metadata-annotated BioSample
source species. The width of the lines is proportional to the number of
runs for each category. For instance, if an sOTU has 10 outgoing
lines, then it appears in 10 different runs. Line colors are determined by
predicted phylum.
NOTE: for large viromes with many sOTUs, this will take a long time to
render and will likely be completely uninterepretable. It is recommended
that you filter by a specific viral phylum.
}
\examples{
con <- palmid::SerratusConnect()
virome <- getVirome(tax = "Meloidogyne", con = con)
drawVirusSankey(virome = virome,
                phylumFilter = c("Pisuviricota", "Kitrinoviricota"),
                colors = c("#E69F00", "#56B4E9"))

}
