% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viromeQueries.R
\name{getVirusTaxonomy}
\alias{getVirusTaxonomy}
\title{getVirusTaxonomy}
\usage{
getVirusTaxonomy(otu = NULL, con = NULL)
}
\arguments{
\item{con}{A connection to the Serratus database}

\item{virus}{A character vector of virus names}

\item{rank}{A character vector of taxonomic ranks}
}
\value{
A character vector of taxons
}
\description{
Return the taxonomic information at a specified rank for
a given virus.
}
