% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palmPrevalence.R
\name{palmPrevalence}
\alias{palmPrevalence}
\title{palmPrevalence}
\usage{
palmPrevalence(virome)
}
\arguments{
\item{virome}{A virome object}
}
\value{
A plotly scatterplot
}
\description{
For every sotu in a virome, plot the number of runs in which
it was found against the number of bioprojects. Points are sized by
normalized mean coverage. Colors are genbank divergence.
}
\examples{
con <- palmid::SerratusConnect()
virome <- getVirome("Meloidogyne", con = con)
palmPrevalence(virome)


}
