% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityCalculations.R
\name{plotAlphaDiversity}
\alias{plotAlphaDiversity}
\title{plotAlphaDiversity}
\usage{
plotAlphaDiversity(virome)
}
\arguments{
\item{virome}{A virome object}
}
\description{
Automatically plots Simpson, Shannon, and Pielou evenness for
a virome object.
}
