% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hostViralBurden.R
\name{plotVirusPositive}
\alias{plotVirusPositive}
\title{plotVirusPositive}
\usage{
plotVirusPositive(vbDF = NULL)
}
\arguments{
\item{vbDF}{viral burden dataframe. See getHostViralBurden()}
}
\value{
A plotly bar chart
}
\description{
Plots the proportion of virus positive runs for each source
species in the virome.
}
\examples{
vbDF <- getHostViralBurden(virome = virome)
plotVirusPositive(vbDF)

}
